Move for- and while-loops to their own submodule
